Refactor attention layer to use half-precision (FP16) for Adam parameters and memory allocation

+++ gpu/attention.h
@@ -8,6 +8,7 @@
 #include <stdbool.h>
 #include <cublasLt.h>
 #include <cuda_runtime.h>
+#include <cuda_fp16.h>
 
 // CUDA Error checking macro
 #ifndef CHECK_CUDA
@@ -62,15 +63,15 @@ typedef struct {
     float* d_W_o_grad; // [d_model x d_model]
     
     // Adam parameters
-    float* d_W_q_m, *d_W_q_v;  // First and second moments for W_q
-    float* d_W_k_m, *d_W_k_v;  // First and second moments for W_k
-    float* d_W_v_m, *d_W_v_v;  // First and second moments for W_v
-    float* d_W_o_m, *d_W_o_v;  // First and second moments for W_o
-    float beta1;               // Exponential decay rate for first moment
-    float beta2;               // Exponential decay rate for second moment
-    float epsilon;             // Small constant for numerical stability
-    int t;                     // Time step
-    float weight_decay;        // Weight decay parameter for AdamW
+    __half* d_W_q_m, *d_W_q_v;  // First and second moments for W_q
+    __half* d_W_k_m, *d_W_k_v;  // First and second moments for W_k
+    __half* d_W_v_m, *d_W_v_v;  // First and second moments for W_v
+    __half* d_W_o_m, *d_W_o_v;  // First and second moments for W_o
+    float beta1;                // Exponential decay rate for first moment
+    float beta2;                // Exponential decay rate for second moment
+    float epsilon;              // Small constant for numerical stability
+    int t;                      // Time step
+    float weight_decay;         // Weight decay parameter for AdamW
     
     // Forward pass buffers
     float* d_Q;            // Query matrix [batch_size x seq_len x d_model]
